refactor: move parseEvent to starknet provider
