+++ src/index.ts
@@ -1,6 +1,5 @@
 import Checkpoint from './checkpoint';
 export { LogLevel } from './utils/logger';
-export { parseEvent } from './utils/events';
 export { AsyncMySqlPool } from './mysql';
 export { createGetLoader } from './graphql';
 export * from './types';
